requirements.txt
Flask
Flask_Bcrypt
Flask_Cors
Flask_JWT_Extended
flask_sqlalchemy
langchain
python-dotenv
tenacity
flask_limiter
mysql-connector-python
gunicorn
flask_migrate